fix: 단어 품사 조회 수정

+++ src/main/java/com/numo/wordapp/service/word/WordGroupService.java
@@ -1,4 +1,4 @@
-package com.numo.wordapp.service;
+package com.numo.wordapp.service.word;
 
 import com.numo.wordapp.comm.exception.CustomException;
 import com.numo.wordapp.comm.exception.ErrorCode;
@@ -6,6 +6,7 @@ import com.numo.wordapp.dto.word.group.WordGroupReadResponseDto;
 import com.numo.wordapp.dto.word.group.WordGroupRequestDto;
 import com.numo.wordapp.dto.word.group.WordGroupResponseDto;
 import com.numo.wordapp.entity.word.detail.WordGroup;
+import com.numo.wordapp.repository.word.WordDetailRepository;
 import com.numo.wordapp.repository.word.group.WordGroupRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class WordGroupService {
     private final WordGroupRepository wordGroupRepository;
+    private final WordDetailRepository wordDetailRepository;
 
     /**
      * 해당하는 유저의 품사 목록을 가져온다.
@@ -30,7 +32,7 @@ public class WordGroupService {
     }
 
     /**
-     * 해당하는 품사 아이디, 유저의 품사 데이터와 detail에 해당하는 데이터도 가져온다
+     * 해당하는 유저, 품사 아이디의 품사 데이터를 가져온다.
      * @param userId 유저 아이디
      * @param wordGroupId 품사 아이디
      * @return 조회한 품사 데이터
@@ -49,9 +51,7 @@ public class WordGroupService {
      */
     public WordGroupResponseDto saveWordGroup(Long userId, WordGroupRequestDto requestDto) {
         // 동일한 품사명은 저장 불가
-        if (wordGroupRepository.existsByUser_UserIdAndName(userId, requestDto.name())) {
-            throw new CustomException(ErrorCode.WORD_GROUP_EXISTS);
-        }
+        checkGroup(userId, requestDto);
         WordGroup wordGroup = requestDto.toEntity(userId);
         return WordGroupResponseDto.of(wordGroupRepository.save(wordGroup));
     }
@@ -65,19 +65,39 @@ public class WordGroupService {
      */
     @Transactional
     public WordGroupResponseDto updateWordGroup(Long userId, Long wordGroupId, WordGroupRequestDto requestDto) {
+        checkDefaultGroup(wordGroupId);
+        // 동일한 품사명은 저장 불가
         WordGroup wordGroup = wordGroupRepository.findWordGroupByIdAndUser(wordGroupId, userId);
+        checkGroup(userId, requestDto);
         wordGroup.update(requestDto);
         return WordGroupResponseDto.of(wordGroup);
     }
 
     /**
-     * 품사 삭제, 품사와 연관되어 있는 데이터 모두 삭제된다.
+     * 품사 삭제, 연관된 품사 상세 데이터가 있으면 삭제하지 않는다.
+     * 기본 품사는 삭제가 불가능
      * @param userId 유저 아이디
      * @param wordGroupId 품사 아이디
      */
     public void removeWordGroup(Long userId, Long wordGroupId) {
+        checkDefaultGroup(wordGroupId);
         WordGroup wordGroup = wordGroupRepository.findWordGroupByIdAndUser(wordGroupId, userId);
-        wordGroup.remove();
+        if (wordDetailRepository.existsByWordGroupAndWord_User_UserId(wordGroup, userId)) {
+            throw new CustomException(ErrorCode.WORD_GROUP_DELETE_FAILED);
+        }
+
         wordGroupRepository.delete(wordGroup);
+    }
+
+    private void checkDefaultGroup(Long wordGroupId) {
+        if (wordGroupRepository.existsByWordGroupIdAndDefaultGroup(wordGroupId, "Y")) {
+            throw new CustomException(ErrorCode.DEFAULT_GROUP_EDIT_FAILED);
+        }
+    }
+
+    private void checkGroup(Long userId, WordGroupRequestDto requestDto) {
+        if (wordGroupRepository.existsGroup(userId, requestDto.name())) {
+            throw new CustomException(ErrorCode.WORD_GROUP_EXISTS);
+        }
     }
 }
